pyproject.toml
[tool.poetry]
name = "sam-example-app"
version = "0.1.0"
description = "AWS SAM experiments"
authors = ["ihoubr <ivo.houbrechts@nike.com>"]
packages = [
    { include = "sam_example_app", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.43"
requests = "^2.25.1"
aws-lambda-powertools = "^1.9.1"
pynamodb = "^5.0.0"
jsonpath-ng = "^1.5.2"
python-dateutil = "^2.8.1"
opentelemetry-api = "^0.17b0"
opentelemetry-propagator-b3 = "^0.17b0"
opentelemetry-sdk = "^0.17b0"
opentelemetry-instrumentation-requests = "^0.17b0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
j2cli = "^0.3.10"
pytest-mock = "^3.5.1"
locust = "^1.4.1"
pytest-dynamodb = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
